+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Removed support for Django 2
+- Now uses the extensions logger
 - Add pre-commit checks to CI
 - Integrate codecov
 - Migrate to allianceauth-app-utils